+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/populator/MagicChunkHandler.java
@@ -3,8 +3,6 @@ package com.elmakers.mine.bukkit.world.populator;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.logging.Level;
-import javax.annotation.Nullable;
 
 import org.bukkit.configuration.ConfigurationSection;
 
@@ -12,7 +10,6 @@ import com.elmakers.mine.bukkit.magic.MagicController;
 import com.elmakers.mine.bukkit.world.MagicWorld;
 
 public class MagicChunkHandler {
-    public static final String BUILTIN_CLASSPATH = "com.elmakers.mine.bukkit.world.populator.builtin";
 
     private final MagicWorld world;
     private final List<MagicBlockPopulator> chunkPopulators = new ArrayList<>();
@@ -21,12 +18,12 @@ public class MagicChunkHandler {
         this.world = world;
     }
 
-    public void load(String worldName, ConfigurationSection config) {
-        if (config == null) return;
-        for (String key : config.getKeys(false)) {
-            ConfigurationSection handlerConfig = config.getConfigurationSection(key);
+    public void load(String worldName, ConfigurationSection populatorConfigs) {
+        if (populatorConfigs == null) return;
+        for (String key : populatorConfigs.getKeys(false)) {
+            ConfigurationSection handlerConfig = populatorConfigs.getConfigurationSection(key);
             if (handlerConfig == null) {
-                getController().getLogger().warning("Was expecting a properties section in world chunk_generate config for key '" + worldName + "', but got: " + config.get(key));
+                getController().getLogger().warning("Was expecting a properties section in world populators config for key '" + worldName + "', but got: " + populatorConfigs.get(key));
                 continue;
             }
             if (!handlerConfig.getBoolean("enabled", true)) {
@@ -34,9 +31,10 @@ public class MagicChunkHandler {
             }
 
             String className = handlerConfig.getString("class");
-            MagicBlockPopulator populator = createBlockPopulator(className);
+            MagicBlockPopulator populator = MagicBlockPopulator.create(getController(), className);
             if (populator != null) {
-                if (populator.load(handlerConfig, world)) {
+                populator.setKey(key);
+                if (populator.load(world, handlerConfig)) {
                     chunkPopulators.add(populator);
                     getController().info("Adding " + key + " populator to " + worldName);
                 } else {
@@ -48,43 +46,21 @@ public class MagicChunkHandler {
         }
     }
 
-    public Collection<MagicBlockPopulator> getPopulators() {
-        return chunkPopulators;
-    }
-
-    @Nullable
-    protected MagicBlockPopulator createBlockPopulator(String className) {
-        if (className == null) return null;
-
-        if (className.indexOf('.') <= 0) {
-            className = BUILTIN_CLASSPATH + "." + className;
-            if (!className.endsWith("Populator")) {
-                className += "Populator";
+    public void load(String worldName, List<String> populators) {
+        if (populators == null || populators.isEmpty()) return;
+        for (String key : populators) {
+            MagicBlockPopulator populator = world.getController().getWorlds().createPopulator(world, key);
+            if (populator != null) {
+                chunkPopulators.add(populator);
+                getController().info("Adding " + key + " populator to " + worldName);
+            } else {
+                getController().info("Skipping invalid " + key + " populator for " + worldName);
             }
         }
+    }
 
-        Class<?> handlerClass = null;
-        try {
-            handlerClass = Class.forName(className);
-        } catch (Throwable ex) {
-            getController().getLogger().log(Level.WARNING, "Error loading chunk populator: " + className, ex);
-            return null;
-        }
-
-        Object newObject;
-        try {
-            newObject = handlerClass.getDeclaredConstructor().newInstance();
-        } catch (Throwable ex) {
-            getController().getLogger().log(Level.WARNING, "Error loading chunk populator: " + className, ex);
-            return null;
-        }
-
-        if (newObject == null || !(newObject instanceof MagicBlockPopulator)) {
-            getController().getLogger().warning("Error loading chunk populator: " + className + ", does it extend MagicChunkPopulator?");
-            return null;
-        }
-
-        return (MagicBlockPopulator)newObject;
+    public Collection<MagicBlockPopulator> getPopulators() {
+        return chunkPopulators;
     }
 
     public boolean isEmpty() {