+++ Magic/CHANGELOG.md
@@ -11,6 +11,9 @@
    - Crops: Generate simple crop farms
    - Sequence: Run multiple other generators in order
    - Random: Randomly choose one generator from a weighted probability map
+ - Create block populators via `populators` configuration
+   - These can use pre-existing populator classes, such as Replace, MagicChest, MagicBlock and Chest
+   - Added new populators: Tower
  - Use generators in world configs for completely custom world gen
  - Allow modifying game rules of a world via `game_rules` config
  - Add `time` parameter to set a world to a specific time of day
@@ -31,6 +34,7 @@
 
  - Changed `arenas.enabled` config option to `enable_arenas`
  - The `chunk_generate` property has been replaced by `populators`, though the former is still supported
+   - This can now be a string list, referring to populators defined in populator configs
  - Removed support for the Elementals plugin (not updated since MC 1.7, I assume no one will miss it)
 
 ## New Features